% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot2pval.R
\name{boot2pval}
\alias{boot2pval}
\title{Bootstrap interval for a two sample test}
\usage{
boot2pval(x1, x2, conf.level = 0.95, iter = 3000, mudiff = 0, test = "two")
}
\arguments{
\item{x1}{sample 1}

\item{x2}{sample 2}

\item{conf.level}{the confidence level, default 0.95}

\item{iter}{times of iteraction for bootstrap}

\item{mudiff}{the difference between mean of population in the Null hypothesis}

\item{test}{indicate two-tailed or one-tailed test, could be "two","uppper","lower"}
}
\value{
the p value and a plot
}
\description{
Bootstrap interval for a two sample test
}
\examples{
set.seed(10);x=rnorm(20,mean=4,sd=2);  y=rnorm(20,mean=4,sd=2)
boot2pval(x1=x,x2=y)
}
