% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateTSS.R
\name{calculateTSS}
\alias{calculateTSS}
\title{Calcuatling TSS}
\usage{
calculateTSS(data)
}
\arguments{
\item{data}{A dataframe containing column 'Height'}
}
\value{
TSS Value
}
\description{
This function is for task 7 on lab 4.
}
\examples{
# calculateTSS(spruce)

}
