% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_beta_mle.R
\name{estimate_beta_mle}
\alias{estimate_beta_mle}
\title{Maximum Likelihood Estimation for Beta Distribution}
\usage{
estimate_beta_mle(x, alpha, beta, ...)
}
\arguments{
\item{x}{A numeric vector containing the sample data.}

\item{alpha}{A numeric vector of possible \eqn{alpha} values.}

\item{beta}{A numeric vector of possible \eqn{beta} values.}

\item{...}{Additional arguments passed to the plot function.}
}
\value{
A list containing the maximum likelihood estimates for \eqn{alpha} and \eqn{beta},
the log-likelihood values, and the plot of the likelihood surface.
}
\description{
This function computes the maximum likelihood estimates for the parameters
\eqn{alpha} and \eqn{beta} of the Beta distribution based on resampled data.
}
