% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findmode.R
\name{findmode}
\alias{findmode}
\title{findmode}
\usage{
findmode(x)
}
\arguments{
\item{x}{A numerical vector}
}
\value{
the mode of the given data set
}
\description{
this function returns the mode of a given data set, which is the value in the data set that occurs the most amount of times
}
\examples{
x=c(1, 2, 3, 3, 3, 5, 6, 7, 7, 8, 9, 11, 12, 13, 13, 13, 13, 17, 20)
mode(x)
result = 13
}
