% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{FIREDAM Dataset}
\format{
A data frame with X rows and Y columns:
\describe{
  \item{DISTANCE}{Distance from the fire station (in miles)}
  \item{DAMAGE}{Amount of fire damage (in thousands of dollars)}
}
}
\source{
<Provide dataset source if available>
}
\usage{
fire
}
\description{
A dataset containing information about fire damage incidents.
}
\examples{
data("fire")
head(fire)
}
\keyword{datasets}
