% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MATH4753Package.R
\docType{package}
\name{Math4753Package}
\alias{Math4753Package}
\title{Aaron Richterkessing package}
\description{
Functions taken from MATH 4753 labs and other projects.
}
\author{
Aaron Richterkessing \email{aaronr019@ou.edu}
}
