% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{my_ci}
\alias{my_ci}
\title{Calculate the confidence interval of a small size of sample}
\usage{
my_ci(d, conf.int = 0.95)
}
\arguments{
\item{d}{data array contains data, the data size <30}

\item{conf.int}{the confidence,default 95%}
}
\value{
a confidence interval
}
\description{
Calculate the confidence interval of a small size of sample
}
\examples{
set.seed(23);x = rnorm(30,mean=10,sd=12)
yourpackage::myci(x)
}
