% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot.R
\name{myboot}
\alias{myboot}
\title{Bootstrap simulation - Lab 9}
\usage{
myboot(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iterations}

\item{x}{sample to be bootstrapped}

\item{fun}{function to be estimated}

\item{alpha}{alpha parameter of  interval}

\item{cx}{point graphical variable}

\item{...}{}
}
\value{
1-alpha confidence interval for a function
}
\description{
Bootstrap simulation - Lab 9
}
\examples{
myboot2(x = rnorm(20, mean = 10, sd = 2), alpha = 0.1, fun = "sd")
}
