% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltb}
\alias{mycltb}
\title{Simulate sample means from a binomial distribution}
\usage{
mycltb(n, iter, p)
}
\arguments{
\item{n}{Sample size}

\item{iter}{Number of iterations}

\item{p}{Probability of success in each trial}
}
\value{
A histogram of the sample means is plotted. Invisibly returns the sample means as a numeric vector.
}
\description{
This function simulates sample means from a binomial distribution
and plots their histogram.
}
\examples{
mycltb(n = 10, iter = 1000, p = 0.5)

}
