% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{mycltp}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{size of the sample}

\item{iter}{number of iterations}

\item{lambda}{value for distribution}

\item{...}{any other specifications}
}
\value{
a histogram of the sample mean, a barplot of sampled y, and a barplot for the probability function of y.
}
\description{
This function creates a Poisson distribution with a random sample
}
\examples{
mycltp(n=2,iter=10000, lambda = 10)
}
