% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myddt.R
\name{myddt}
\alias{myddt}
\title{Project 1 - myddt()}
\usage{
myddt(df, species)
}
\arguments{
\item{df}{RIVER, MILE, SPECIES, LENGTH, WEIGHT, DDT}

\item{species}{CCATFISH, SMBUFFALO, LMBASS}
}
\value{
A plot of Length vs Weight for a Species of fish
}
\description{
Project 1 - myddt()
}
\examples{
myddt(df = ddt, SPECIES  = "CCATFISH")
}
