% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Max Likelihood Function}
\usage{
mymaxlik(lfun, x, param)
}
\arguments{
\item{lfun}{the type of function}

\item{x}{the data vector}

\item{param}{probability parameter}
}
\value{
a maximum likelihood estimate with a corresponding graph
}
\description{
Max Likelihood Function
}
\examples{
\dontrun{y <- c(3,3,4,3,4,5,5,4);logbin=function(x,param) log(dbinom(x,prob=param,size=10));mymaxlik(x=y,param=seq(0,1,length=1000),lfun=logbin)}
}
