% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlnorm.R
\name{mymlnorm}
\alias{mymlnorm}
\title{Max likelihood estimation of mu and sigma for normal distribution}
\usage{
mymlnorm(x, mu, sig, ...)
}
\arguments{
\item{x}{sample data}

\item{mu}{a sequence of possible mu}

\item{sig}{a sequence of possible sigma}

\item{...}{}
}
\value{
a plot containing the max value of likelihood with corresponding mu and sigma
}
\description{
Max likelihood estimation of mu and sigma for normal distribution
}
\examples{
mymlnorm(x=c(5,7,7,8,10),mu=seq(5,10,length=1000),sig=seq(0.1,4,length=1000),lwd=2,labcex=1)
}
