% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myNegBin.R
\name{myNegBin}
\alias{myNegBin}
\title{Negative Binomial Probability Distribution Function}
\usage{
myNegBin(y, r, p)
}
\arguments{
\item{y}{quantile to find as a vector}

\item{r}{size as a vector}

\item{p}{probability as a vector}
}
\value{
a negative binomial probability distribution
}
\description{
Negative Binomial Probability Distribution Function
}
\examples{
\dontrun{myNegBin(10,3,0.4)}
}
