% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myNRML.R
\name{myNRML}
\alias{myNRML}
\title{Newton-Raphson}
\usage{
myNRML(x0, delta = 0.001, llik, xrange, parameter = "param")
}
\arguments{
\item{x0}{initial root guess}

\item{delta}{acceptable difference}

\item{llik}{likelihood function}

\item{xrange}{sets the range of x}

\item{parameter}{sets the parameter}
}
\value{
root estimated by Newton Raphson estimation
}
\description{
Newton-Raphson
}
\examples{
myNRML(x0=1,delta=0.000001,llik=function(x) log(dpois(12,x)*dpois(10,x)),xrange=c(0,20),parameter="lambda" )
}
