% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myplot.R
\name{myplot}
\alias{myplot}
\title{make a function of fitted result for quad.lm}
\usage{
myplot(x)
}
\arguments{
\item{x}{a number, when it is not setted, the function return a function could be used in curve() to draw the fitted line.}
}
\value{
a number predicted by the quad.lm or used in curve()
}
\description{
make a function of fitted result for quad.lm
}
\examples{
curve(myplot, lwd=2, col="steelblue")
}
