% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypvalue.R
\name{mypvalue}
\alias{mypvalue}
\title{mypvalue}
\usage{
mypvalue(t0, xmax = 4, n = 20, alpha = 0.05)
}
\arguments{
\item{t0}{this is tcalc for the data set}

\item{xmax}{the maximum x value for the plot}

\item{n}{the sample size for the data set}

\item{alpha}{The variable that determines the confidence interval}
}
\value{

}
\description{
FUNCTION for P-values: Display P-value areas
}
\examples{
set.seed(55);x1=rnorm(30,mean=25,sd=5)
mypvalue((mean(x1)-24)/(sd(x1)/sqrt(30)),n=30,alpha=0.05)
}
