% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myread.R
\name{myread}
\alias{myread}
\title{Read csv file into table frame}
\usage{
myread(csv)
}
\arguments{
\item{csv}{a csv file in a defined dird}
}
\value{
a table frame
}
\description{
Read csv file into table frame
}
\examples{
dird="C:\\\\Users\\\\stew9983\\\\OneDrive - University of Oklahoma\\\\DATAxls\\\\";
mpg.df=myread("EPAGAS.csv")
}
