% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysample.R
\name{mysample}
\alias{mysample}
\title{Generate a Sample and Create a Bar Plot}
\usage{
mysample(n, iter = 1, time = 1)
}
\arguments{
\item{n}{Sample size.}

\item{iter}{Number of iterations (default: 1).}

\item{time}{Pause time between iterations in seconds (default: 1).}
}
\value{
A bar plot of relative frequencies.
}
\description{
This function randomly selects numbers from 1 to 10 with replacement
and creates a bar plot showing the relative frequencies.
}
\examples{
SPRING254753nguy0700::mysample(n = 1000, iter = 30)
}
