% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets function}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{= Number of Plane Seats}

\item{gamma}{= Probability of Overbooking}

\item{p}{= Probability of Not Showing}
}
\value{
mylist, plotd, plotc
}
\description{
ntickets function
}
\examples{
ntickets(N=200, gamma=0.02, p=0.95)
}
