% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterhist.R
\name{scatterhist}
\alias{scatterhist}
\title{Scatter-Hist}
\usage{
scatterhist(x, y, xlab = "", ylab = "")
}
\arguments{
\item{ylab}{}
}
\value{
A sophisticated plot
}
\description{
Scatter-Hist
}
\examples{
\dontrunwith{(ddt, scatterhist(LENGTH,WEIGHT, xlab="LENGTH"))}
}
