% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqandsqrt.R
\name{sqandsqrt}
\alias{sqandsqrt}
\title{Square Root Function}
\usage{
sqandsqrt(x)
}
\arguments{
\item{x}{}
}
\value{
a list of squares and their respective roots
}
\description{
Takes a list of numbers and calculates both their square and square roots.
}
\examples{
x = 4, then s = 16, and sr = 4
}
