% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{FIREDAM data set}
\format{
A data frame with 16 rows and 2 variables:
\describe{
\item{DISTANCE}{distance in miles}
\item{DAMAGE}{damage scale of fire}
...
}
}
\usage{
fire
}
\description{
Data collected about distances and damages of fires
}
\keyword{datasets}
