% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire Damage Data}
\format{
A data frame with X rows and 2 variables:
\describe{
  \item{DISTANCE}{Distance to nearest fire station (in miles)}
  \item{DAMAGE}{Amount of fire damage (in thousands of dollars)}
}
}
\source{
Collected by a fire insurance company, as described in *Statistics for Engineering and the Sciences* (6th ed.)
}
\usage{
fire
}
\description{
A dataset containing observations of fire damage and distance from fire stations.
}
\keyword{datasets}
