% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.r
\docType{data}
\name{fire}
\alias{fire}
\title{Fire Damage and Distance from the Fire Station}
\format{A data frame with 15 rows and 2 variables:
\describe{
  \item{damage}{	numeric	Damage (\$000s)}
  \item{distance}{	numeric	Distance from the fire station (miles)}
  ...
}}
\source{
\url{no}
}
\usage{
fire
}
\description{
House damage and distance from the fire station, of 15 house fires. Data collected by an insurance company for homes in a particular area.
}
\keyword{datasets}
