% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Mean Bootstrapping Function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5)
}
\arguments{
\item{iter}{the amount of iterations}

\item{x}{the sample}

\item{fun}{the type of function}

\item{alpha}{1 minus the confidence}

\item{cx}{the cx value}
}
\value{
a histogram of the bootstrapped sample and confidence interval
}
\description{
Mean Bootstrapping Function
}
\examples{
\dontrun{myboot2(x=rnorm(20,mean=10,sd=4))}
}
