% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{myboot2}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iterations}

\item{x}{population to be sampled from}

\item{fun}{function being applied to the data}

\item{alpha}{1-alpha is the confidence interval}

\item{cx}{character expansion factor}

\item{...}{}
}
\value{
Histogram of bootstrap sample statistics
}
\description{
This function creates a confidence interval using bootstrap methods
}
\examples{
myboot2(iter=10000, x=rnorm(30, 0, 1), alpha = 0.05, fun = "mean", cx=1.5)
}
