% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrapping function}
\usage{
myboot2(x, fun, nboot = 1000, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{fun}{A function to apply to each bootstrap sample.}

\item{nboot}{Number of bootstrap samples. Default is 1000.}

\item{...}{Additional arguments passed to \code{fun}.}
}
\value{
A vector of bootstrap estimates.
}
\description{
Performs bootstrap sampling of a statistic from a numeric vector.
}
\examples{
x <- rnorm(100)
boot_means <- myboot2(x, mean)
hist(boot_means)
}
