% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{Simlulate the Central limit theory using Poisson distribution}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{the sample size of each cycle}

\item{iter}{the number of cycle}

\item{lambda}{the parameters of Poisson distribution}

\item{...}{}
}
\value{
histogram of simulating and theoretic distribution
}
\description{
Simlulate the Central limit theory using Poisson distribution
}
\examples{
windows();mycltp(n=10,iter=10000,lambda=5)
}
