% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Normal Function Graph}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{the mean}

\item{sigma}{the standard deviation}

\item{a}{the upper bound of area}
}
\value{
a normal function graph with the given parameters and probability of the desired area
}
\description{
Normal Function Graph
}
\examples{
\dontrun{myncurve(10,5,6)}
}
