% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{plot the normal density and shade the area of P(x<a)}
\usage{
myncurve(mu, sigma, a = a)
}
\arguments{
\item{mu}{the mean of normal distribution}

\item{sigma}{the sd of normal distribution}

\item{a}{for calulate the P(x<a)}
}
\value{
a plot contain the normal density and shading area, also, the probability of x< a is also printed
}
\description{
plot the normal density and shade the area of P(x<a)
}
\examples{
mycurve(1,2,a=3)
}
