% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Tickets to be Sold}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{the number of seats in a flight}

\item{gamma}{the probability a plane will be truly overbooked}

\item{p}{the probability of a "show"}
}
\value{
The number of tickets to be sold for a plane
}
\description{
Tickets to be Sold
}
\examples{
\dontrun{ntickets(200,0.02,0.95)}
}
