% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterhist.R
\name{scatterhist}
\alias{scatterhist}
\title{co-plot the data using scatter and histogram}
\usage{
scatterhist(x, y, xlab = "", ylab = "")
}
\arguments{
\item{x}{the variable as x coordinate}

\item{y}{the variable as y coordinate}

\item{xlab}{the name of x axis}

\item{ylab}{the name of y axis}
}
\value{
a plot containing both points and histogram
}
\description{
co-plot the data using scatter and histogram
}
\examples{
ddt=read.csv("../lab1/DDT.csv");
with(ddt, scatterhist(LENGTH,WEIGHT, xlab="LENGTH"))
}
